v0.0.3beta code readability
