+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "videotrimmer",
-    "version": "0.0.2",
+    "version": "0.0.3",
     "license": "MIT",
     "description": "easy video trimmer",
     "author": {
